Fix bugs and warnings
1. Use explicit aligned and unaligned loading of 128-bit vector
   registers for the MT19937.
2. Remove comparisons between signed and unsigned integers.
3. Use memcpy to obey strict aliasing rules.

+++ pxart/pxart/mt19937.hpp
@@ -49,7 +49,7 @@ struct mt19937 {
   constexpr result_type max() noexcept { return (~uint_type{}) & mask; }
 
   uint_type state[state_size];
-  int state_index = state_size;
+  size_t state_index = state_size;
 };
 
 struct mt19937::default_seeder {
@@ -77,15 +77,15 @@ constexpr mt19937::mt19937() : mt19937{default_seeder{}} {}
 
 constexpr mt19937::result_type mt19937::operator()() noexcept {
   if (state_index >= state_size) {
-    const auto transition = [this](int k, int k1, int k2) constexpr {
+    const auto transition = [this](size_t k, size_t k1, size_t k2) constexpr {
       const auto x = (state[k] & upper_mask) | (state[k1] & lower_mask);
       // state[k] = state[k2] ^ (x >> 1) ^ ((x & 0x01) * xor_mask);
       state[k] = state[k2] ^ (x >> 1) ^ ((state[k1] & 0x01) * xor_mask);
     };
 
-    for (int k = 0; k < state_size - shift_size; ++k)
+    for (size_t k = 0; k < state_size - shift_size; ++k)
       transition(k, k + 1, k + shift_size);
-    for (int k = state_size - shift_size; k < state_size - 1; ++k)
+    for (size_t k = state_size - shift_size; k < state_size - 1; ++k)
       transition(k, k + 1, k + shift_size - state_size);
     transition(state_size - 1, 0, shift_size - 1);
